requirements.txt
aws-lambda-powertools==2.26.1
awswrangler==3.4.1
boto3==1.28.84
pre-commit==3.5.0
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.7.0
virtualenv==20.24.6
